pyproject.toml
[tool.poetry]
name = "madness-deblender"
version = "v1.0_beta1"
description = "Galaxy deblender from variational autoencoders"
authors = ["Biswajit Biswas <biswas@apc.in2p3.fr>"]
readme = "README.md"
keywords = ["galaxies", "deblending", "cosmology", "generative models"]
license = "MIT"
include = ["madness_deblender/data/*", "btksims"]
repository = "https://github.com/b-biswas/madness"

[tool.poetry.dependencies]
python = "^3.8"
numpy = "*"
pandas = "*"
astropy = "*"
tensorflow = "2.11.0"
tensorflow-probability = "0.19.0"
tensorflow-datasets = "*"
scikit-image = "*"
sep = "*"
galcheat = "*"
numba = "*"
btksims = {path="btksims", optional=true, develop=true}
matplotlib = {version="*", optional=true}
seaborn = {version="*", optional=true}
pytest = {version="*", optional=true}
pre-commit = {version="*", optional=true}
jupyter = {version="*", optional=true}
ipykernel = {version="*", optional=true}
autograd = {version="*", optional=true}
proxmin = {version="*", optional=true}
pybind11 = {version="*", optional=true}


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
sims = ["btksims"]
notebooks = ["jupyter", "ipykernel", "btksims", "matplotlib", "seaborn"]
dev = ["pytest", "pre-commit", "jupyter", "ipykernel", "btksims", "matplotlib", "seaborn"]
